Fixes `timespan_in_words` in single-unit mode returning a seconds part.
If it converted the specified seconds to e.g. hours, there could be a
fraction of a second left over due to floating point errors. That
fraction of a second would then be part of the output string.

Fix entails only making the chosen unit part of the output string.